+++ src/main/java/com/codecrafters/companity/adapter/post/dto/request/RequestForUpdatingPost.java
@@ -1,6 +1,6 @@
 package com.codecrafters.companity.adapter.post.dto.request;
 
-import com.codecrafters.companity.adapter.post.mapper.PostMapperForController;
+import com.codecrafters.companity.adapter.post.mapper.PostMapper;
 import com.codecrafters.companity.domain.enums.City;
 import com.codecrafters.companity.domain.enums.Sport;
 import com.codecrafters.companity.domain.post.PostForUpdate;
@@ -17,6 +17,6 @@ public class RequestForUpdatingPost {
     private String content;
 
     public PostForUpdate toPostUpdateDto(){
-        return PostMapperForController.POST_MAPPER.toPostForUpdate(this);
+        return PostMapper.POST_MAPPER.toPostForUpdate(this);
     }
 }
